requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
imutils==0.5.4
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.1.5
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
tifffile==2020.9.3
